Movie classes to their own files
